Improved root motion extraction; separated file with root motion extraction data
