+++ SoulsFormats/Formats/HKX/HavokAnimationData.cs
@@ -51,225 +51,6 @@ namespace SoulsFormats.Havok
         }
     }
 
-
-    public class NewRootMotionHandlerData
-    {
-        //public Matrix CurrentAbsoluteRootMotion = Matrix.Identity;
-
-        public readonly Vector4 Up;
-        public readonly Vector4 Forward;
-        public readonly float Duration;
-        public readonly Vector4[] Frames;
-
-        /// <summary>
-        /// The accumulative root motion delta applied by playing the entire anim from the beginning to the end.
-        /// </summary>
-        public readonly Vector4 LoopDeltaForward;
-
-        /// <summary>
-        /// The accumulative root motion delta applied by playing the entire anim in reverse from the end to the beginning.
-        /// </summary>
-        public readonly Vector4 LoopDeltaBackward;
-
-        public NewRootMotionHandlerData(HKX.HKADefaultAnimatedReferenceFrame refFrame) : this(refFrame.Up, refFrame.Forward, refFrame.Duration, refFrame.ReferenceFrameSamples.GetArrayData().Elements.Select(hkxVector=>hkxVector.Vector).ToArray())
-        {
-
-        }
-
-        public NewRootMotionHandlerData(Vector4 up, Vector4 forward, float duration, Vector4[] frames)
-        {
-            Up = up;
-            Forward = forward;
-            Duration = duration;
-            Frames = frames;
-
-            LoopDeltaForward = frames[frames.Length - 1] - frames[0];
-            LoopDeltaBackward = frames[0] - frames[frames.Length - 1];
-        }
-
-
-        // public bool Accumulate = true;
-
-        private Matrix4x4 GetMatrixFromSample(Vector4 sample)
-        {
-            return Matrix4x4.CreateRotationY(sample.W) *
-                Matrix4x4.CreateWorld(
-                    new Vector3(sample.X, sample.Y, sample.Z),
-                    new Vector3(Forward.X, Forward.Y, -Forward.Z),
-                    new Vector3(Up.X, Up.Y, Up.Z));
-        }
-
-        public Vector4 GetSample(float time)
-        {
-            float frame = Frames.Length * time / Duration;
-
-            float frameFloor = (float)Math.Floor(frame % (Frames.Length - 1));
-            Vector4 sample = Frames[(int)frameFloor];
-
-            if (frame != frameFloor)
-            {
-                float frameMod = frame % 1;
-
-                Vector4 nextFrameRootMotion;
-
-                //if (frame >= Frames.Length - 1)
-                //    nextFrameRootMotion = Frames[0];
-                //else
-                //    nextFrameRootMotion = Frames[(int)(frameFloor + 1)];
-
-                nextFrameRootMotion = Frames[(int)(frameFloor + 1)];
-
-                sample = Vector4.Lerp(sample, nextFrameRootMotion, frameMod);
-            }
-
-            return sample;
-        }
-
-        private static Vector4 AddRootMotion(Vector4 start, Vector4 toAdd, float direction, bool dontAddRotation = false)
-        {
-            if (!dontAddRotation)
-                start.W += toAdd.W;
-
-            Vector3 displacement = Vector3.Transform(new Vector3(toAdd.X, toAdd.Y, toAdd.Z), Matrix4x4.CreateRotationY(direction));
-            start.X += displacement.X;
-            start.Y += displacement.Y;
-            start.Z += displacement.Z;
-            return start;
-        }
-
-        public (Vector3 positionChange, float directionChange) ExtractRootMotion(float previousTime, float nextTime)
-        {
-            int fullLoops = (int)( (nextTime - previousTime) / (Duration));
-
-            if (nextTime > Duration  && fullLoops == 0)
-                {
-                fullLoops = 1;
-            }
-
-
-            Vector4 fullLoopRootMotion = Math.Abs(fullLoops) *( fullLoops > 0 ? LoopDeltaForward : LoopDeltaBackward);
-
-
-            float startingFrame = ((previousTime % Duration) + Duration) % Duration;
-
-            float finalFrame = ((nextTime  % Duration) + Duration) % Duration;
-
-            var previousData = GetSample(startingFrame);
-
-            var nextData = GetSample(finalFrame);
-
-
-            Vector4 extractedRootMotion = nextData - previousData + fullLoopRootMotion;
-
-            return (new Vector3(extractedRootMotion.X, extractedRootMotion.Y, extractedRootMotion.Z), extractedRootMotion.W);
-        }
-
-        public (Vector4 Motion, float Direction) UpdateRootMotion(Vector4 currentRootMotion, float currentDirection, float currentFrame, int loopCountDelta)
-        {
-            var currentData = GetSample(currentFrame);
-
-            var changedRootMotion = AddRootMotion(currentRootMotion, currentData, currentDirection);
-
-            return (changedRootMotion, currentDirection);
-        }
-
-        public (Vector4 Motion, float Direction) UpdateRootMotion(Vector4 currentRootMotion, float currentDirection, float previousFrame, float currentFrame, int loopCountDelta)
-        {
-            var currentData = GetSample(currentFrame);
-
-            var previousData = GetSample(previousFrame);
-            //currentRootMotion
-            var changedRootMotion = AddRootMotion(Vector4.Zero, currentData - previousData, currentDirection);
-
-            while (loopCountDelta != 0)
-            {
-                if (loopCountDelta > 0)
-                {
-                    changedRootMotion = AddRootMotion(changedRootMotion, LoopDeltaForward, currentDirection, true);
-                    currentDirection += LoopDeltaForward.W;
-                    loopCountDelta--;
-                }
-                else if (loopCountDelta < 0)
-                {
-                    changedRootMotion = AddRootMotion(changedRootMotion, LoopDeltaBackward, currentDirection, true);
-                    currentDirection += LoopDeltaBackward.W;
-                    loopCountDelta++;
-                }
-            }
-
-            float rotationChange = changedRootMotion.W;
-
-            Vector4 hackSomething = currentRootMotion;
-
-            hackSomething.W = 0;
-
-            changedRootMotion.W = 0;
-
-            Vector4 appliedRootMotion = hackSomething + changedRootMotion;
-
-            return (Motion: appliedRootMotion, Direction: currentDirection + rotationChange);
-        }
-
-        //public (Vector4 Motion, float Direction) UpdateRootMotion(Vector4 currentRootMotion, Vector4 prevFrameData, float currentDirection, float currentFrame, int loopCountDelta)
-        //{
-        //    var nextFrameData = GetSample(currentFrame);
-
-        //    Vector4 targetRootMotion;
-
-        //    if (Accumulate)
-        //    {
-        //        //currentRootMotion *= GetMatrixFromSample(nextFrameData - prevFrameData);
-
-        //        Vector4 changedRootMotion = AddRootMotion(currentRootMotion, nextFrameData - prevFrameData, currentDirection, dontAddRotation: true);
-
-        //        changedRootMotion.W = nextFrameData.W;
-
-        //        while (loopCountDelta != 0)
-        //        {
-        //            if (loopCountDelta > 0)
-        //            {
-        //                changedRootMotion = AddRootMotion(changedRootMotion, LoopDeltaForward, currentDirection, dontAddRotation: true);
-        //                currentDirection += LoopDeltaForward.W;
-        //                loopCountDelta--;
-        //            }
-        //            else if (loopCountDelta < 0)
-        //            {
-        //                changedRootMotion = AddRootMotion(changedRootMotion, LoopDeltaBackward, currentDirection, dontAddRotation: true);
-        //                currentDirection += LoopDeltaBackward.W;
-        //                loopCountDelta++;
-        //            }
-        //        }
-
-        //        targetRootMotion = changedRootMotion;
-        //    }
-        //    else
-        //    {
-        //        currentDirection = 0;
-        //        targetRootMotion = AddRootMotion(Vector4.Zero, nextFrameData, currentDirection);
-        //    }
-
-        //    // TODO: This code remains of what was in DSAnimStudio.
-        //    // Maybe remove it
-        //    // prevFrameData = nextFrameData;
-
-        //    //CurrentAbsoluteRootMotion = GetMatrixFromSample(GetSample(currentFrame));
-
-        //    //if (forceAbsoluteRootMotion)
-        //    //{
-
-        //    //}
-        //    //else
-        //    //{
-        //    //    CurrentAbsoluteRootMotion *= GetMatrixFromSample(nextFrameData - prevFrameData);
-        //    //}
-
-
-
-        //    return (targetRootMotion, currentDirection);
-        //}
-    }
-
-
     public abstract class HavokAnimationData
     {
         public HKX.AnimationBlendHint BlendHint = HKX.AnimationBlendHint.NORMAL;
